Situation-Assessment/WP6_AIBAST/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.6.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.0
idna==3.10
ipython==8.37.0
jedi==0.19.2
jinja2==3.1.4
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-community==0.3.17
langchain-core==0.3.36
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.6
langgraph==0.3.0
langgraph-checkpoint==2.0.16
langgraph-prebuilt==0.1.0
langgraph-sdk==0.1.53
langsmith==0.3.8
markupsafe==2.1.5
marshmallow==3.26.1
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
neo4j==5.28.1
networkx==3.3
numpy==1.26.4
openai==1.97.1
orjson==3.10.15
packaging==24.2
parso==0.8.4
pillow==11.0.0
prompt-toolkit==3.0.51
propcache==0.2.1
pure-eval==0.2.3
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.7.1
pygments==2.19.2
python-dotenv==1.0.1
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
sniffio==1.3.1
sqlalchemy==2.0.38
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing-extensions==4.12.2
typing-inspect==0.9.0
urllib3==2.3.0
wcwidth==0.2.13
yarl==1.18.3
zstandard==0.23.0
